adding fuction declare case and rturn case to delcrlist and statlist, implementing canendparse for ending main parse process

+++ lexical.h
@@ -51,7 +51,7 @@
 
 typedef union{
     int integerNumber;
-    double doubleNumber;
+    float doubleNumber;
     string *str;
 }tokenContent;
